lambda/requirements.txt
boto3==1.35.70
botocore==1.35.70
jmespath==1.0.1
pillow==11.0.0
python-dateutil==2.9.0.post0
s3transfer==0.10.4
six==1.16.0
urllib3==2.2.3
